More edits

+++ BackEnd/src/main/java/com/CraftyCorner/Capstone/Common/UserRepository.java
@@ -2,14 +2,13 @@ package com.CraftyCorner.Capstone.Common;
 
 
 import com.CraftyCorner.Capstone.Model.MyAppUser;
-import com.CraftyCorner.Capstone.Model.MyAppUser;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends JpaRepository<MyAppUser, Integer> {
+public interface UserRepository extends JpaRepository<MyAppUser, Long> {
     MyAppUser findByUsername(String username);
     MyAppUser findByEmail(String email);
+
 
 }
